Make HistoryAwareReferenceField consistent with versioning
If no version is explicitly set for the target object in the instance,
assume the instance points to the very first version of that object.
Note that this only happens when an Analysis is associated to a
calculation that hasn't been edited yet. If the analysis was associated
to an already versioned caculation, the analysis itself will always
store the exact version.