Display the version of the calcultion used in analysis info
The Analysis info view (popup) displays the calculation bound to the
Analysis Service by default, rather than the calculation bound to the
analysis object. Since the relationship between an analysis object and
a calculation is governed by a reference history aware field, the
calculation displayed might not be the same that is actually used on
results sumbmission.

This commit gives preference to the analysis over the service on
calculation retrieval and displays version information too.