Give priority to Analysis over Service when displaying Calculation information (#1261)
* Traceback on AR creation with Calculation never edited

When one of the Analysis to be created within an AnalysisRequest
is bound to a Calculation that hasn't been never edited after its
creation, a Traceback arises:

2019-03-05 11:12:12 ERROR Zope.SiteErrorLog 1551780732.540.735467473409 http://localhost:8080$
Traceback (innermost last):
  Module ZPublisher.Publish, line 138, in publish
  Module ZPublisher.mapply, line 77, in mapply
  Module ZPublisher.Publish, line 48, in call_object
  Module bika.lims.browser.analysisrequest.add2, line 48, in decorator
  Module bika.lims.browser.analysisrequest.add2, line 707, in __call__
  Module bika.lims.browser.analysisrequest.add2, line 1606, in ajax_submit
  Module bika.lims.utils.analysisrequest, line 70, in create_analysisrequest
  Module Products.Archetypes.BaseObject, line 633, in processForm
  Module Products.Archetypes.BaseObject, line 624, in _processForm
   - __traceback_info__: (<AnalysisRequest at /senaite/clients/client-1/16546a274e024c64debf8$
  Module Products.Archetypes.utils, line 130, in mapply
  Module Products.Archetypes.ClassGen, line 76, in generatedMutator
  Module bika.lims.browser.fields.aranalysesfield, line 148, in set
  Module bika.lims.utils.analysis, line 83, in create_analysis
  Module bika.lims.utils.analysis, line 68, in copy_analysis_field_values
  Module Products.Archetypes.ClassGen, line 76, in generatedMutator
  Module bika.lims.browser.fields.historyawarereferencefield, line 88, in set
  Module Products.CMFEditions.CopyModifyMergeRepositoryTool, line 298, in save
  Module Products.CMFEditions.CopyModifyMergeRepositoryTool, line 431, in _assertAuthorized
Unauthorized: You are not allowed to access 'save' in this context

The reason is that historyawarereferencefield's setter tries to find
if the current version of the target object (Calculation to be assigned
to the Analysis object) is the latest. If the function finds out that
there isn't any version yet for the target object, the function tries to
create the first version automatically here:

```python
version_id = hasattr(targets[uid], 'version_id') and \
                     targets[uid].version_id or None
if version_id is None:
    # attempt initial save of unversioned targets
    pr = getToolByName(instance, 'portal_repository')
    if pr.isVersionable(targets[uid]):
        pr.save(obj=targets[uid],
                comment=to_utf8(ts(_("Initial revision"))))
```

For this action to be possible, the current user must have the permission
"CMFEditions: Save new version" granted for the target object (in this
case, a Calcultion), which is not always true: e.g. client contacts do not
have privileges for setup items other than "Access contents information",
"List folder contents" and "View".

This commit removes this code from historyawarereferncefield

* Make HistoryAwareReferenceField consistent with versioning

If no version is explicitly set for the target object in the instance,
assume the instance points to the very first version of that object.
Note that this only happens when an Analysis is associated to a
calculation that hasn't been edited yet. If the analysis was associated
to an already versioned caculation, the analysis itself will always
store the exact version.

* Create the first version (0) of Calculation on save

* Display the version of the calcultion used in analysis info

The Analysis info view (popup) displays the calculation bound to the
Analysis Service by default, rather than the calculation bound to the
analysis object. Since the relationship between an analysis object and
a calculation is governed by a reference history aware field, the
calculation displayed might not be the same that is actually used on
results sumbmission.

This commit gives preference to the analysis over the service on
calculation retrieval and displays version information too.

* Removed event subscriber for calculations

* Removed at_post_create_script for calculations

* Removed unused code