Fix AttributeError: state_variable
Make the upgrade step more robust. Found a "Batch" object, in a database with
42,906 batches, that was bound to a workflow for which the "state_variable"
was inexplicably missing. Because of the same reason, it was not possible to
retrieve the review_history of that object.

With this change, we safely handle these edge cases, but since we rebind a
new workflow later and create a new initial_state if not available, we actually
fix the problem.

Traceback (innermost last):
  Module ZPublisher.Publish, line 138, in publish
  Module ZPublisher.mapply, line 77, in mapply
  Module ZPublisher.Publish, line 48, in call_object
  Module Products.GenericSetup.tool, line 1053, in manage_doUpgrades
  Module Products.GenericSetup.upgrade, line 166, in doStep
  Module bika.lims.upgrade, line 68, in wrap_func_args
  Module bika.lims.upgrade.v01_03_000, line 245, in upgrade
  Module bika.lims.upgrade.v01_03_000, line 685, in update_workflows
  Module bika.lims.upgrade.v01_03_000, line 2280, in restore_review_history_for_affected_objects
  Module bika.lims.upgrade.v01_03_000, line 2293, in restore_review_history_for
  Module bika.lims.upgrade.v01_03_000, line 2343, in get_purged_review_history_for
  Module bika.lims.upgrade.v01_03_000, line 2406, in create_initial_review_history
AttributeError: state_variable