Use id in api.get_id if a brain and id attr
The api's function relies on "getId" metadata only atm. Nevertheless,
uid_catalog is oftenly used to retrieve brains and this catalog does
not contain the index getId. One option would be to add this same
index in uid_catalog, but since getId is a function that overrides
Plone.Archetypes.BaseObject.getId, that in turn does `return self.id`,
we can assume that getId will always return the same value as `id`.

The motivation of this change is that when a huge database is used and
the id formatting for a given content type (say AnalysisRequest) is
changed, the reseed of the id causes the instance to wake-up all objects
stored. If the amount of objects is ~100k or more, it means that the
instance gets freezed and the RAM consumption easily increases to values
between 8G and 10G. Of course, this causes a huge foodprint on performance.

With this change, the system behaves faster in this scenario and the RAM
is kept under control to a reasonable value of ~4G max (db with >1M objects).