No need to look for siblings if calc does not have services
Function getDependencies is called each time an analysis is submitted.
Dependencies are those analyses the current analysis depends on based
on the calculation. getDependencies was calling getSiblings even when
calculation did not have dependent services.

This commit makes the function faster for those analyses bound to
a calculation that do not depend on other analyses.