Slightly better performance when retrieving analysis dependencies (#1363)
* No need to look for siblings if calc does not have services

Function getDependencies is called each time an analysis is submitted.
Dependencies are those analyses the current analysis depends on based
on the calculation. getDependencies was calling getSiblings even when
calculation did not have dependent services.

This commit makes the function faster for those analyses bound to
a calculation that do not depend on other analyses.

* Remove unnecessary code in worksheet listing

This code makes the listing to be rendered slowly because it inspects
the analyses included in every single worksheet and eventually wakes
up the container, mostly AnalysisRequest objects. The interesting part
is that this code is not useful at all.

* Remove unnecessary column getLayout from worksheets listing catalog

This metadata was only used to make the worksheets list slower

* Changelog

* Better text

* Log the time it takes any given transition

* Typo

* replace not service_uids by len(service_uids) == 0