Fix unauthorized access (#1394)
Security info for analyses from inside samples that have been labeled
for internal use needs to be updated. Otherwise when a client contact
user access to a primary sample with an internal-use partition, the
system throws an Unauthorized message because new security seetings are
not yet taken into consideration when querying against analysis catalog
in order to retrieve the analyses from child partitions.