Simplify analysis.isOpen() and make it faster
The possible statuses before an analysis is submitted are "registered",
"assigned" and "unassigned". In earlier versions, isOpen was checking
the review status for "assigned" and "unassigned", but isOpen function
was used everywhere together with isRegistered. Once an analysis is
submitted, the object is marked with ISubmitted interface.

This commit omits the check for review status, rather it only checks
if the object is provided by ISubmitted interface, that makes the
system to behave faster and reduces the complexity in those places where
both isOpen and isRegistered where checked.