Supersed analysis.isOpen() by not ISubmitted.providedBy(analysis) (#1405)
* Simplify analysis.isOpen() and make it faster

The possible statuses before an analysis is submitted are "registered",
"assigned" and "unassigned". In earlier versions, isOpen was checking
the review status for "assigned" and "unassigned", but isOpen function
was used everywhere together with isRegistered. Once an analysis is
submitted, the object is marked with ISubmitted interface.

This commit omits the check for review status, rather it only checks
if the object is provided by ISubmitted interface, that makes the
system to behave faster and reduces the complexity in those places where
both isOpen and isRegistered where checked.

* Remove isRegistered function

This function is no longer used because of the previous commit

* Removal of analysis.isOpen()

With this commit, the function isOpen has been removed in favour of
direct check for ISubmitted marker interface.

* Typo