Use the value set for ui_item property when displaying ReferenceWidget (#1424)
* Notify user with failing addresses when emailing of results reports

* Changelog

* Handle unicode and translation

* Use the value set for ui_item property when displaying ReferenceWidget

When reference widget is rendered by default for a field with a value already
set (before user selects an item from the list), the value displayed in the
input field is always the Title of the referenced object. This is fine in
most cases, except those for which the the value for `ui_item` differs from
'Title'. In such case, the field will display the value for the attr that
matches with `ui_item` property after an item is selected, but not the first
time is rendered (or when displayed in read-only mode).