Added support for adapters in guard handler
This commit allows to adapt the behavior of core's guard
handler. Adapters that implement IGuardAdapter have at
least one of the two functions below:

pre_guard(action)
This function is executed first and the core's guard will
only be evaluated if the return of pre_guard is True.

post_guard(action)
This function is only executed when the core's guard
returns True.

Therefore, the outcome of guard_handler will depend on the
following order: pre_guard > core's guard > post_guard