Allow LabClerk to access results view when Worksheet is verified
The permissions for manage_results view were wrong. EditWorksheet
permission makes sense in views that are used for adding analyses,
etc. (e.g. AddDuplicates, Add Control, etc.), but the permission
for analyses listing (manage_results) in Worksheet should be
govened by ViewResults only.

If the current status of the analysis does not allow its edition
for the current user, this is handled automatically by folderitems,
so there is no sense to apply the permission for EditWorksheet at
this lsiting level.