Fix jsonapi.read omits `include_methods` when a single parameter is used (#1493)
* Fix get_include_methods function from jsonapi

When include_methods param is used in a read call with a single
parameter (instead of a list of more than one), the jsonapi does
not handle the value parameter properly and transforms it to
a list of characters instead.

A call like follows:

```
this.ajax_submit({
  url: this.get_portal_url() + "/@@API/read",
  data: {
          catalog_name: "uid_catalog",
          UID: widget.attr('data-uid'),
          include_methods: ["getSomething"],
        }
}).done(function(data) {
  return deferred.resolveWith(this, [data.objects[0]["getSomething"]]);
});
```

fails. The function `get_include_methods` returns `getSomething`
instead of a list `[getSomething]`. Therefore, the function
`load_method_values` tries the following calls to the instance:
`g`, `e`, `t`, `S`, `o`, ...

If the call is like follows:
```
this.ajax_submit({
  url: this.get_portal_url() + "/@@API/read",
  data: {
          catalog_name: "uid_catalog",
          UID: widget.attr('data-uid'),
          include_methods: ["getSomething", "getId"],
        }
}).done(function(data) {
  return deferred.resolveWith(this, [data.objects[0]["getSomething"]]);
});
```
the list is resolved properly.

* Now we are here, make it a bit better

* Changelog

Co-authored-by: Ramon Bartl <rb@ridingbytes.com>