Ensure analysis' result range keeps all range subfields
When creating a Sample with "Sample Specifications" setting enabled,
the result ranges assigned to the analyses are not fully compliant
with those coming from the selected specification (if any), cause
they only keep min_warn, min, max and max_warn.

This commit ensures that when a Sample is created, the result ranges
from analyses are fully compliant with the subfields defined in the
specifications, even if these subfields come from different add-ons
(e.g. from senaite.panic).