Cannot reject Sample when Contact has no email set
When "Notify on Sample Rejection" option is enabled in Setup, the system
sends an e-mail with the rejection reasons to the contact assigned to the
sample. If the sample neither has a contact with a valid email nor CCEmail
is empty, the system fails to reject because of empty recipient.

```
Traceback (innermost last):
  Module ZPublisher.Publish, line 138, in publish
  Module ZPublisher.mapply, line 77, in mapply
  Module ZPublisher.Publish, line 48, in call_object
  Module bika.lims.browser.analysisrequest.add2, line 66, in decorator
  Module bika.lims.browser.analysisrequest.add2, line 734, in __call__
  Module bika.lims.browser.analysisrequest.add2, line 1676, in ajax_submit
  Module bika.lims.utils.analysisrequest, line 146, in create_analysisrequest
  Module bika.lims.utils.analysisrequest, line 503, in do_rejection
  Module bika.lims.api.mail, line 231, in send_email
  Module Products.MailHost.MailHost, line 231, in send
  Module Products.MailHost.MailHost, line 450, in _mungeHeaders
MailHostError: No message recipients designated
```