Unable to get the previous status when duplicated in review history (#1572)
The function `get_prev_status_from_history` gets the review history
of a given object, do a reverse of the list and then keeps iterating
through the fetched status until a match against the passed-in value
is found. In such case, it returns the next status (since the list
was reversed, it actually returns the status reached by the object
previously in time). Nevertheless, if the status is repeated in the
list consecutively, the function always return the same passed-in
status.

This commit address the situation, so even if the object has a given
status consecutively duplicated in its review_history, the function
returns the status the object had before