Do not force available workflow transitions in batches listing
The batches listing defines the allowed transitions to be
performed against selected batches when a given filter is selected.

Available transitions for any given object are bound to its workflow
definition and current user privileges. Forcing the available
transitions makes other add-ons with the batch workflow customized to
also tweak these values for new transitions to be displayed.