Allow to retest analyses without the need of retraction (#1582)
* Add "retest" transition for analysis

At the moment, retests can only be created by retracting the
analysis. This is fine for when the analysis had a wrong result
and the labman wants to discard the result and the analyst to
perform the same test again.

Nevertheless, for some tests, a confirmative test is required
when the first result, although "valid", is considered as
non-conclusive. In such case, the "retract" transition is
missleading.

A "retest" transition has been added and only available when
the analysis is in "to_be_verified" status. When triggered, the
current analysis is transitioned to "verified" status, but a
retest is created at the same time. Therefore, the sample
ends-up with two "valid" analyses for same keyword.

* Change func signatures (replace param retracted by with_retests)

* Change "Retested" (column header) by "Retest"

* Changelog

* Fix doctests massive failure (omg)

* Fix test

* Guarantee dependents/dependencies in all hierarchy are retested

* Cleanup

* Always display retests below the original analysis in listings