Removal of "manage_results" view from Sample (#1587)
* Remove redirect to manage_results for AnalysisRequest

This commit removes the redirection to sample's manage results view when the
sample is in received status and user has Edit privileges. Currently, when these
conditions are met, the manage results view is displayed, but this view is
always displayed, even when "View" link is clicked. For sake of simplicity,
better to handle this in e.g. listing rather than modifying the expression url
for AnalysisRequest's action view.

* Remove redirect to manage_results (AR) after barcode

* Remove nuts from AR "manage_analyses" req

* Fix barcode_entry tests

* Remove AnalysisRequestManageResultsView