Fix interims with choices are not displayed in listings after a while
When a value is set to an interim with choices, the listing initially
displays the value correctly because remains in `senaite.app.listing`
cache, but as soon as the list is filtered or the page refreshed, the
value is not displayed anymore. Listing grabs the interims from the
brain directly (`getInterimFields` metadata column), but even that
`set_field` endpoint from `senaite.app.listing` stores the value and
reindexes the analysis object, the value of the metadata field remains
unchanged.

This metadata column is only used in analysis listing and we already
wake-up the analysis object to grab other properties while rendering
the list, so there is no advantage at all of having this metadata
column.

Also, the fact that interims field is stored as a dict requires
special care when storing/updating this value in ZODB. Again, because
of the limited use of it, the complexity it entails is not worth.