Fix RecordsWidget does not store hidden fields in Add form (#1771)
When RecordsWidget containing a ReferenceWidgets inside is used
in Sample Add form, the hidden fields (typically _uid and those
defined in the colModel) are not set for the given row when the
referenced object is selected.

The reason is that the old-fashion js did not take into
consideration that in Sample Add form, the (sample) column is
injected into the fieldnames. This made the js to not be able to
calculate the row number properly and therefore, not able to
set the values for hidden fields

This commit assumes grabs the last digit of the fieldname to
get the rownumber instead of picking the value at position 3.