Remove ACTIONS_TO_INDEXES mapping
When an action takes place for a given object, the system tries
to reindex only those indexes that "matter". However, this action-indexes
mapping is harcoded as a dictionary and even if it wasn't, the problems
with data integrity it entails is not worth. Quite often, values from
indexes in catalogs are not updated in accordance because of this and
therefore, their value does not represent the real value stored in the
object. E.g. "modified".