Make the prepublish guard to rely on interfaces instead of statuses
Relying on interfaces is better, especially when there are other
add-ons that extend the analyses workflow with other statuses and
transitions.

For instance, one might have a custom transition that moves the
analysis to the status `out_of_stock`, but we do want the system to
behave as if it was a verified analysis. With the current
implementation we don't have a choice other than overwriting the
guards to allow the user to pre-publish and publish samples with
one single analysis (in `out_of_stock` status).

With this PR, one can simply mark the object with the
proper interface `IVerified` in an `AfterTransitionEvent` and the
system will behave as expected without the need of monkey patching
the guards from core.